Refonte engagement/statistiques : enums forts, tests, XML
- Migration des modules d’engagement/statistiques vers FranceJudo.Metier.ExtensionNoyau
- Passage généralisé aux enums typés pour les états/types (remplacement des int)
- Adaptation complète des entités, méthodes métier et sérialisation XML
- Ajout et adaptation des tests unitaires pour la nouvelle architecture
- Ajout de CompetitionType2Enum, corrections sur la gestion de la casse et le logger
- Suppressions de règles de style et ajustements projet/tests

+++ KernelImplTests/Noyau/Deroulement/VueCombatTests.cs
@@ -70,7 +70,7 @@ namespace KernelImpl.Tests.Noyau.Deroulement
             Mock<IOrganisationData> mockOrganisation = new Mock<IOrganisationData>();
             Mock<ICompetition> mockCompetition = new Mock<ICompetition>();
 
-            mockCompetition.Setup(c => c.type).Returns( (int) CompetitionTypeEnum.Equipe);
+            mockCompetition.Setup(c => c.type).Returns( CompetitionTypeEnum.Equipe);
             mockOrganisation.SetupGet(o => o.Competition).Returns(mockCompetition.Object);
             mockDc.SetupGet(d => d.Organisation).Returns(mockOrganisation.Object);
             mockDc.Setup(dc => dc.Participants).Returns(mockParticipants.Object);
